requirements.txt
# Fine-tune Once, Serve Anywhere - Project Dependencies
# ====================================================
# NLP II Final Project - TinyLlama Fine-tuning Comparison
# 
# Installation:
#   1. Create venv: python -m venv venv
#   2. Activate: source venv/bin/activate
#   3. Install PyTorch: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#   4. Install all dependencies: pip install -r requirements.txt
#
# Note: All dependencies including Unsloth are now included in this file.

# =====================================================
# Core Deep Learning (Install separately for CUDA)
# =====================================================
# torch>=2.1.0
# torchvision>=0.16.0
# torchaudio>=2.1.0
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# =====================================================
# Transformers & PEFT (QLoRA)
# =====================================================
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.4
flash-attn>=2.5.0

# =====================================================
# Unsloth (Optimized Fine-tuning)
# =====================================================
unsloth @ git+https://github.com/unslothai/unsloth.git
unsloth_zoo
xformers

# =====================================================
# vLLM (High-throughput Inference)
# =====================================================
vllm>=0.2.7

# =====================================================
# Ollama (Local Inference)
# =====================================================
# Install Ollama separately: https://ollama.ai/download
ollama>=0.1.6

# =====================================================
# Dataset & Data Processing
# =====================================================
datasets>=2.15.0
pandas>=2.0.0
numpy>=1.24.0

# =====================================================
# Evaluation Metrics
# =====================================================
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0

# =====================================================
# Visualization & Analysis
# =====================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
kaleido>=0.2.1

# =====================================================
# Configuration & Utilities
# =====================================================
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0

# =====================================================
# Memory & Performance Monitoring
# =====================================================
psutil>=5.9.0
gputil>=1.4.0
nvidia-ml-py>=12.535.0

# =====================================================
# Development Tools
# =====================================================
jupyter>=1.0.0
ipykernel>=6.27.0
ipywidgets>=8.1.0
